Use fnet_char_t * for strings
Strings are arrays of char (and thus fnet_char_t) in C and the
compiler may complain if they are unsigned. Also "XX" is placed
in flash and is therefore of type const char.

It isn't an issue unless enabling debug information. So this patch
updates the cases in the debug and trace output.

Also, UDP port numbers goes up to 65535, which is 5 digits. So
change the output to also align properly for numbers above 999.